+++ NEWS.md
@@ -6,6 +6,7 @@
 ## Minor changes
 * Added `show.html` to `MotifDiscovery` and `MotifCompare` to control the pop-up window.
 * Fixed bug in `InferRegulator`.
+* Fixed bug in `DEbCA` (output ChIP-seq and ATAC-seq peak information).
 
 -------------